requirements.txt
azure-core==1.30.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
isodate==0.6.1
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
